charm-dns-service/requirements.txt
ops ~= 2.5
python-openstackclient
